+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/repository/ProcessDBQueryRepository.scala
@@ -5,7 +5,7 @@ import pl.touk.nussknacker.engine.api.process.{ProcessId, ProcessName, ScenarioV
 import pl.touk.nussknacker.security.Permission
 import pl.touk.nussknacker.ui.db.NuTables
 import pl.touk.nussknacker.ui.db.entity._
-import pl.touk.nussknacker.ui.security.api.{AdminUser, CommonUser, LoggedUser}
+import pl.touk.nussknacker.ui.security.api.{AdminUser, CommonUser, ImpersonatedUser, LoggedUser, RealLoggedUser}
 import pl.touk.nussknacker.ui.{BadRequestError, NotFoundError}
 
 import java.sql.Timestamp
@@ -18,9 +18,15 @@ trait ProcessDBQueryRepository[F[_]] extends Repository[F] with NuTables {
   protected def processTableFilteredByUser(
       implicit loggedUser: LoggedUser
   ): Query[ProcessEntityFactory#ProcessEntity, ProcessEntityData, Seq] = {
+    def getTableForUser(user: RealLoggedUser) = {
+      user match {
+        case user: CommonUser => processesTable.filter(_.processCategory inSet user.categories(Permission.Read))
+        case _: AdminUser     => processesTable
+      }
+    }
     loggedUser match {
-      case user: CommonUser => processesTable.filter(_.processCategory inSet user.categories(Permission.Read))
-      case _: AdminUser     => processesTable
+      case user: RealLoggedUser   => getTableForUser(user)
+      case user: ImpersonatedUser => getTableForUser(user.impersonatedUser)
     }
   }