+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/repository/ProcessRepository.scala
@@ -18,7 +18,7 @@ import pl.touk.nussknacker.ui.process.repository.ProcessRepository.{
   ProcessUpdated,
   UpdateProcessAction
 }
-import pl.touk.nussknacker.ui.security.api.LoggedUser
+import pl.touk.nussknacker.ui.security.api.{ImpersonatedUser, LoggedUser, RealLoggedUser}
 import slick.dbio.DBIOAction
 
 import java.sql.Timestamp
@@ -112,7 +112,9 @@ class DBProcessRepository(protected val dbRef: DbRef, modelVersion: ProcessingTy
       isFragment = action.isFragment,
       isArchived = false,
       createdAt = Timestamp.from(now),
-      createdBy = userName
+      createdBy = userName,
+      impersonatedByIdentity = loggedUser.impersonatingUserId,
+      impersonatedByUsername = loggedUser.impersonatingUserName
     )
 
     val insertNew =