+++ extensions-api/src/main/scala/pl/touk/nussknacker/security/AuthCredentials.scala
@@ -5,6 +5,14 @@ sealed trait AuthCredentials
 object AuthCredentials {
 
   final case class PassedAuthCredentials(value: String) extends AuthCredentials
-  case object AnonymousAccess                           extends AuthCredentials
+
+  final case class ImpersonatedAuthCredentials(
+      impersonatingUserCredentials: PassedAuthCredentials,
+      impersonatedUserIdentity: ImpersonatedUserIdentity
+  ) extends AuthCredentials
+
+  case object NoCredentialsProvided extends AuthCredentials
 
 }
+
+final case class ImpersonatedUserIdentity(value: String)